backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
cryptography==43.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
python-multipart==0.0.9